src/esql/operations/discover.py
from __future__ import annotations

from empire_commons import list_util
from efs.path import find_parent_sibling_from_path, join, get_file_name_no_extension
from efs.file_system import remake_dir, create_file, mkdir
from efs.io_ import OpenFileForWrite

